improve moleculer files
