+++ template/server/index.js
@@ -1,6 +1,6 @@
 import express from 'express'
 import { Nuxt, Builder } from 'nuxt'
-import service from './service';
+import moleculer from './moleculer';
 
 const app = express()
 const host = process.env.HOST || '127.0.0.1'
@@ -9,7 +9,7 @@ const port = process.env.PORT || 3000
 app.set('port', port)
 
 // Import API Routes
-app.use('/api', service.express());
+app.use('/api', moleculer.express());
 
 // Import and Set Nuxt.js options
 let config = require('../nuxt.config.js')
